+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @unipasswallet/sdk
 
+## 0.3.1-alpha.1
+
+### Patch Changes
+
+- @unipasswallet/provider: add session key support
+- Updated dependencies
+  - @unipasswallet/abi@0.3.1-alpha.1
+  - @unipasswallet/dkim-base@0.3.1-alpha.1
+  - @unipasswallet/keys@0.3.1-alpha.1
+  - @unipasswallet/network@0.3.1-alpha.1
+  - @unipasswallet/relayer@0.3.1-alpha.1
+  - @unipasswallet/transaction-builders@0.3.1-alpha.1
+  - @unipasswallet/transactions@0.3.1-alpha.1
+  - @unipasswallet/utils@0.3.1-alpha.1
+  - @unipasswallet/wallet@0.3.1-alpha.1
+
 ## 0.3.1-alpha.0
 
 ### Patch Changes